Add console serial socket
One can use socat to log in via this serial socket if getty is running
on ttyS0.  The biggest benefit, though, is that when an Alpine qemu VM
that was installed from the standard alpine-virt ISO, the VM's syslog no
longer gets filled with errors from getty trying to open the (otherwise
non-existent) /dev/ttyS0.